Factor out indenting of rendered bodies
